Refactored multi agents to be just robots.
