+++ Python/tdw/robot_data/non_moving.py
